pyenv phpenv tmux installer for Debian

+++ setup/tmux.bash
@@ -1,12 +1,21 @@
 #/bin/bash
-
-DEBIAN=(
-    tmux
-)
-
-# Debian
-[ -n "`/usr/bin/apt-get`" ] && sudo apt-get update && sudo apt-get install "${DEBIAN[@]}" -y
+source `dirname $0`/conf.bash
+BIN_INIT tmux
 
 # .tmux.conf
+[ -f ~/bin/home/.tmux.conf  ] || ln -s ~/bin/home/.tmux.conf ~/.tmux.conf;
 
-ln -s ~/bin/home/.tmux.conf ~/.tmux.conf
+if [ -n "`which tmux`"  ]; then
+    BIN_EXIT;
+fi
+
+case "$BIN_OS" in 
+  "DEBIAN") 
+    PKGS=(tmux);;
+ "CENTOS")   
+    PKGS=(tmux);;
+ "OSX")
+    PKGS=(tmux);;
+esac;
+
+$BIN_PKG "${PKGS[@]}";
